requirements.txt
aniso8601==9.0.1
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
commonmark==0.9.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
exceptiongroup==1.0.4
executing==1.2.0
filelock==3.8.2
Flask==2.2.2
httpie==3.2.1
idna==3.4
importlib-metadata==5.0.0
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonify==0.5
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
multidict==6.0.2
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.0
pluggy==1.0.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.2.0
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
requests-toolbelt==0.10.1
rich==12.6.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.0
tabulate==0.9.0
toml==0.10.2
tomli==2.0.1
traitlets==5.5.0
urllib3==1.26.12
virtualenv==20.17.1
virtualenv-clone==0.5.7
wcwidth==0.2.5
Werkzeug==2.2.2
yoyo-migrations==8.0.0
zipp==3.10.0
